Don't allow orphaned admins in the first place
