+++ app/models/admin.rb
@@ -62,10 +62,6 @@ class Admin < ActiveRecord::Base
       end
     end
     alias_method :add_privilege, :add_privileges
-
-    def destroy_orphaned
-      destroy_all '(select count(*) from people where people.admin_id = admins.id and people.deleted = 0) = 0 and template_name is null'
-    end
   end
 
   add_privileges *%w(
